finish 04-hw

+++ src/controllers/notesController.js
@@ -2,10 +2,11 @@ import { Note } from '../models/note.js';
 import createHttpError from 'http-errors';
 
 export const getAllNotes = async (req, res) => {
+  const { _id: userId } = req.user;
   const { page, perPage, sortBy, sortOrder, tag, search } = req.query;
   const skip = (page - 1) * perPage;
 
-  const notesQuery = Note.find();
+  const notesQuery = Note.find({ userId });
 
   if (tag) {
     notesQuery.where('tag').equals(tag);
@@ -36,7 +37,8 @@ export const getAllNotes = async (req, res) => {
 
 export const getNoteById = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findById(noteId);
+  const { _id: userId } = req.user;
+  const note = await Note.findById({ noteId, userId });
 
   if (!note) {
     throw createHttpError(404, 'Note not found');
@@ -46,14 +48,16 @@ export const getNoteById = async (req, res, next) => {
 };
 
 export const createNote = async (req, res) => {
-  const result = await Note.create(req.body);
+  const { _id: userId } = req.user;
+  const result = await Note.create({ ...req.body, userId });
 
   res.status(201).json(result);
 };
 
 export const deleteNote = async (req, res) => {
   const { noteId } = req.params;
-  const result = await Note.findByIdAndDelete(noteId);
+  const { _id: userId } = req.user;
+  const result = await Note.findByIdAndDelete({ noteId, userId });
   if (!result) throw createHttpError(404, 'Note not found');
 
   res.json(result);
@@ -61,7 +65,8 @@ export const deleteNote = async (req, res) => {
 
 export const updateNote = async (req, res) => {
   const { noteId } = req.params;
-  const result = await Note.findByIdAndUpdate(noteId, req.body, {
+  const { _id: userId } = req.user;
+  const result = await Note.findByIdAndUpdate({ noteId, userId }, req.body, {
     new: true,
   });
   if (!result) throw createHttpError(404, 'Note not found');
